requirements.txt
protobuf>=4.21.0,<6.0.0
fastapi==0.111.0
uvicorn[standard]==0.29.0
grpcio>=1.58.0
grpcio-tools>=1.58.0
paho-mqtt==1.6.1